reorg(math) moved area function to math utilities adjusted dependent code
