+++ src/math/utils/index.js
@@ -1,4 +1,5 @@
 module.exports = {
+  area: require('./area'),
   clamp: require('./clamp'),
   degToRad: require('./degToRad'),
   intersect: require('./intersect'),
